+++ changelog.md
@@ -8,3 +8,4 @@
 * Added Travis-CI support
 * Worker has almost identical API as 1.1.1 PECL gearman library
 * Worker no longer returns json as a result (you can send binary data)
+* Client uses same server adding API as 1.1.1 PECL gearman library